requirements.txt
# Core dependencies
streamlit>=1.28.0
groq>=0.4.0
python-dotenv>=1.0.0

# Optional: Development dependencies
# Add these if you want to contribute to the project
# black>=23.0.0
# flake8>=6.0.0
# pytest>=7.0.0
